LPS-151626 Move CommonStylesManager to the correct folder
